Refactor T&T API to use modular router, change root path to '/api'
Moves Track and Trade API to an separate express.Router instance,
and changes the root path from '/tnt', which is nonstandard and
redundant, to '/api'.
